fix: change filesize type from float to int64 in searchable array

+++ config/scout.php
@@ -394,7 +394,7 @@ return [
                         ],
                         [
                             'name' => 'filesize',
-                            'type' => 'float',
+                            'type' => 'int64',
                             'sort' => true,
                         ],
                         [
